+++ android/src/main/java/com/getcapacitor/community/media/MediaPlugin.java
@@ -4,7 +4,6 @@ import android.Manifest;
 import android.app.DownloadManager;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.SystemClock;
@@ -50,9 +49,6 @@ public class MediaPlugin extends Plugin {
 
     private static final String PERMISSION_DENIED_ERROR = "Unable to access media, user denied permission request";
 
-    private static final String MOVIES = "MOVIES";
-    private static final String PICTURES = "PICTURES";
-
     private static final Integer API_LEVEL_29 = 29;
 
     // @todo
@@ -79,7 +75,7 @@ public class MediaPlugin extends Plugin {
         Log.d("DEBUG LOG", "SAVE PHOTO TO ALBUM");
         if (isStoragePermissionGranted()) {
             Log.d("DEBUG LOG", "HAS PERMISSION");
-            _saveMedia(call, PICTURES);
+            _saveMedia(call);
         } else {
             Log.d("DEBUG LOG", "NOT ALLOWED");
             this.bridge.saveCall(call);
@@ -93,7 +89,7 @@ public class MediaPlugin extends Plugin {
         Log.d("DEBUG LOG", "SAVE VIDEO TO ALBUM");
         if (isStoragePermissionGranted()) {
             Log.d("DEBUG LOG", "HAS PERMISSION");
-            _saveMedia(call, MOVIES);
+            _saveMedia(call);
         } else {
             Log.d("DEBUG LOG", "NOT ALLOWED");
             this.bridge.saveCall(call);
@@ -106,7 +102,7 @@ public class MediaPlugin extends Plugin {
         Log.d("DEBUG LOG", "SAVE GIF TO ALBUM");
         if (isStoragePermissionGranted()) {
             Log.d("DEBUG LOG", "HAS PERMISSION");
-            _saveMedia(call, PICTURES);
+            _saveMedia(call);
         } else {
             Log.d("DEBUG LOG", "NOT ALLOWED");
             this.bridge.saveCall(call);
@@ -136,24 +132,10 @@ public class MediaPlugin extends Plugin {
         }
 
         switch (call.getMethodName()) {
-            case "getMedias":
-                call.unimplemented();
-                break;
-            case "getAlbums":
-                _getAlbums(call);
-                break;
-            case "savePhoto":
-                _saveMedia(call, "PICTURES");
-                break;
-            case "saveVideo":
-                _saveMedia(call, "MOVIES");
-                break;
-            case "saveGif":
-                _saveMedia(call, "PICTURES");
-                break;
-            case "createAlbum":
-                _createAlbum(call);
-                break;
+            case "getMedias" -> call.unimplemented();
+            case "getAlbums" -> _getAlbums(call);
+            case "savePhoto", "saveVideo", "saveGif" -> _saveMedia(call);
+            case "createAlbum" -> _createAlbum(call);
         }
     }
 
@@ -169,8 +151,9 @@ public class MediaPlugin extends Plugin {
         Set<String> bucketIds = new HashSet<String>();
 
         String[] projection = new String[] {
-          MediaStore.MediaColumns.BUCKET_DISPLAY_NAME,
-          MediaStore.MediaColumns.BUCKET_ID
+            MediaStore.MediaColumns.BUCKET_DISPLAY_NAME,
+            MediaStore.MediaColumns.BUCKET_ID,
+            MediaStore.MediaColumns.DATA
         };
         Cursor cur = getActivity().getContentResolver().query(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, projection, null, null, null);
 
@@ -178,13 +161,27 @@ public class MediaPlugin extends Plugin {
             String albumName = cur.getString((cur.getColumnIndex(MediaStore.MediaColumns.BUCKET_DISPLAY_NAME)));
             String bucketId = cur.getString((cur.getColumnIndex(MediaStore.MediaColumns.BUCKET_ID)));
 
-            if (!bucketIds.contains(bucketId)){
-              JSObject album = new JSObject();
+            if (!bucketIds.contains(bucketId)) {
+                String path = cur.getString((cur.getColumnIndex(MediaStore.MediaColumns.DATA)));
+                File fileForPath = new File(path);
+                JSObject album = new JSObject();
 
-              album.put("name", albumName);
-              albums.put(album);
+                album.put("name", albumName);
+                album.put("identifier", fileForPath.getParent());
+                albums.put(album);
 
-              bucketIds.add(bucketId);
+                bucketIds.add(bucketId);
+            }
+        }
+
+        File albumPath = new File(getAlbumPath());
+        for (File sub : albumPath.listFiles()) {
+            if (sub.isDirectory() && sub.listFiles().length == 0) {
+                JSObject album = new JSObject();
+
+                album.put("name", sub.getName());
+                album.put("identifier", sub.getAbsolutePath());
+                albums.put(album);
             }
         }
 
@@ -196,14 +193,15 @@ public class MediaPlugin extends Plugin {
         call.resolve(response);
     }
 
-    private void _saveMedia(PluginCall call, String destination) {
-        String dest;
-        if (destination == MOVIES) {
-            dest = Environment.DIRECTORY_MOVIES;
+    private String getAlbumPath() {
+        if (Build.VERSION.SDK_INT >= API_LEVEL_29) {
+            return getContext().getExternalMediaDirs()[0].getAbsolutePath();
         } else {
-            dest = Environment.DIRECTORY_PICTURES;
+            return Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES).getAbsolutePath();
         }
+    }
 
+    private void _saveMedia(PluginCall call) {
         Log.d("DEBUG LOG", "___SAVE MEDIA TO ALBUM");
         String inputPath = call.getString("path");
         if (inputPath == null) {
@@ -256,23 +254,20 @@ public class MediaPlugin extends Plugin {
             inputFile = new File(inputUri.getPath());
         }
 
-        String album = call.getString("album");
+        String album = call.getString("albumIdentifier");
         File albumDir = null;
-        String albumPath;
         Log.d("SDK BUILD VERSION", String.valueOf(Build.VERSION.SDK_INT));
 
-        if (Build.VERSION.SDK_INT >= API_LEVEL_29) {
-            albumPath = getContext().getExternalMediaDirs()[0].getAbsolutePath();
+        if (album != null) {
+            albumDir = new File(album);
         } else {
-            albumPath = Environment.getExternalStoragePublicDirectory(dest).getAbsolutePath();
+            call.error("Album identifier required");
+            return;
         }
 
-        // Log.d("ENV LOG", String.valueOf(getContext().getExternalMediaDirs()));
-
-        if (album != null) {
-            albumDir = new File(albumPath, album);
-        } else {
-            call.error("album name required");
+        if (!albumDir.exists() || !albumDir.isDirectory()) {
+            call.error("Album identifier does not exist, use getAlbums() to get");
+            return;
         }
 
         Log.d("ENV LOG - ALBUM DIR", String.valueOf(albumDir));
@@ -292,17 +287,13 @@ public class MediaPlugin extends Plugin {
     private void _createAlbum(PluginCall call) {
         Log.d("DEBUG LOG", "___CREATE ALBUM");
         String folderName = call.getString("name");
-        String folder;
 
-        if (Build.VERSION.SDK_INT >= 29) {
-            folder = getContext().getExternalMediaDirs()[0].getAbsolutePath() + "/" + folderName;
-        } else {
-            folder = Environment.getExternalStoragePublicDirectory(folderName).toString();
+        if (folderName == null) {
+            call.reject("Album name must be given!");
+            return;
         }
 
-        Log.d("ENV STORAGE", folder);
-
-        File f = new File(folder);
+        File f = new File(getAlbumPath(), folderName);
 
         if (!f.exists()) {
             if (!f.mkdir()) {